Add configuration option to boot an alternative 2nd stage
Add the ability for shim to load an optional configuration file.  This
new file is called "options.csv".  The configuration file is completely
optional. If used, it is located in the same directory as the booted
shim.  The "options.csv" file currently allows a single entry.  Other
options could be added to it in the future.  The first and only entry in
the file is the name of the secondary boot loader shim will load. The
"options.csv" file is in Unicode LE format.

This allows a signed shim to directly load a UKI without the need to
rename it to grub. Shim's transitive trust is maintained. If the
alternative 2nd stage can not be verified, it will not boot.

Signed-off-by: Eric Snowberg <eric.snowberg@oracle.com>